moev component to another file
